refactor(filid): split rule-engine into single-fn files
Split rule-engine.ts (7 top-level fn + 8 inline Rule.check methods) into
per-function files. rule-engine.ts becomes a facade re-exporting the five
public functions, preserving the import path used by core/rules, core
barrel, mcp/tools (rule-query, structure-validate), hooks
(user-prompt-submit), and fractal-validator.

- utils/ gains 8 check-*.ts (one per built-in rule). checkZeroPeerFile is
  a factory over the project's additional-allowed config.
- utils/ gains remap-severity.ts and wrap-exempt.ts for the applyOverrides
  composition chain.
- rule-engine/ root gains apply-overrides.ts, load-builtin-rules.ts,
  get-active-rules.ts, evaluate-rule.ts, evaluate-rules.ts.
