+++ packages/filid/src/core/rules/rule-engine/rule-engine.ts
@@ -1,432 +1,14 @@
 /**
  * @file rule-engine.ts
- * @description 내장 규칙 로드 및 평가 엔진.
+ * @description Facade — re-exports 8개 내장 FCA 규칙 로더 및 평가기.
  *
- * 8개의 내장 규칙을 제공하고 FractalTree에 대해 규칙을 평가한다.
- * loadBuiltinRules(overrides?)를 통해 .filid/config.json 기반 오버라이드를 지원한다.
+ * 내부 파일은 각각 단일 함수를 담고 있으며, 외부 호출자는 이 파일을 통해
+ * 공개 API(loadBuiltinRules, applyOverrides, evaluateRule, evaluateRules,
+ * getActiveRules)에 접근한다. 8개 개별 rule check 구현은 utils/check-*.ts
+ * 에 분리되어 있다.
  */
-import type { FractalTree } from '../../../types/fractal.js';
-import type {
-  Rule,
-  RuleContext,
-  RuleEvaluationResult,
-  RuleOverride,
-  RuleSeverity,
-  RuleViolation,
-} from '../../../types/rules.js';
-import { BUILTIN_RULE_IDS } from '../../../constants/builtin-rule-ids.js';
-import type { ScanOptions } from '../../../types/scan.js';
-import { DEFAULT_SCAN_OPTIONS } from '../../../constants/scan-defaults.js';
-import { ALLOWED_FRACTAL_ROOT_FILES } from '../../../constants/allowed-peer-files.js';
-import type { AllowedEntry } from '../../infra/config-loader/loaders/config-schemas.js';
-import { isExempt } from './utils/is-exempt.js';
-import { isValidNaming } from './utils/is-valid-naming.js';
-
-/**
- * Remap every violation's severity to `severity`. `innerCheck` is whatever
- * check function is currently on the rule — may itself be a wrapper.
- */
-function remapSeverity(
-  innerCheck: Rule['check'],
-  severity: RuleSeverity,
-): Rule['check'] {
-  return (ctx: RuleContext): RuleViolation[] =>
-    innerCheck(ctx).map((v) => ({ ...v, severity }));
-}
-
-/**
- * Short-circuit the check when the node matches any exempt glob. `innerCheck`
- * may itself be a severity-remap wrapper from {@link remapSeverity} — that is
- * intended; exempt wraps whatever check the rule currently exposes, so the
- * outermost wrapper runs first and skips the inner pipeline on a match.
- */
-function wrapExempt(
-  innerCheck: Rule['check'],
-  exempt: string[],
-): Rule['check'] {
-  return (ctx: RuleContext): RuleViolation[] =>
-    isExempt(ctx.node, exempt) ? [] : innerCheck(ctx);
-}
-
-/**
- * 프로젝트별 오버라이드를 내장 규칙에 적용한다.
- *
- * Composes check wrappers in this order: `rule.check` → (optional) severity
- * remap → (optional) exempt wrap. The linear chain makes the sequential
- * composition explicit — each wrapper takes the previous check as input and
- * returns the next one. Rule bodies never see either mechanism, preserving
- * `Rule.check` purity.
- */
-export function applyOverrides(
-  rules: Rule[],
-  overrides: Record<string, RuleOverride>,
-): Rule[] {
-  return rules.map((rule) => {
-    const override = overrides[rule.id];
-    if (!override) return rule;
-    const newEnabled = override.enabled ?? rule.enabled;
-    const newSeverity = override.severity ?? rule.severity;
-    const hasExempt = !!override.exempt && override.exempt.length > 0;
-    if (
-      newEnabled === rule.enabled &&
-      newSeverity === rule.severity &&
-      !hasExempt
-    ) {
-      return rule;
-    }
-    let check = rule.check;
-    if (newSeverity !== rule.severity) check = remapSeverity(check, newSeverity);
-    if (hasExempt) check = wrapExempt(check, override.exempt!);
-    return { ...rule, enabled: newEnabled, severity: newSeverity, check };
-  });
-}
-
-/**
- * 8개 내장 규칙 인스턴스를 생성하여 반환한다.
- * overrides가 전달되면 enabled/severity/exempt를 프로젝트별로 재정의한다.
- * additionalAllowed의 객체형 엔트리는 `paths` glob이 현재 노드 경로와 일치할 때만 basename을 허용한다.
- */
-export function loadBuiltinRules(
-  overrides?: Record<string, RuleOverride>,
-  additionalAllowed?: AllowedEntry[],
-): Rule[] {
-  const rules: Rule[] = [
-    // 1. naming-convention: 디렉토리명이 kebab-case 또는 camelCase여야 한다
-    {
-      id: BUILTIN_RULE_IDS.NAMING_CONVENTION,
-      name: 'Naming Convention',
-      description:
-        'Directory and file names must follow kebab-case or camelCase.',
-      category: 'naming',
-      severity: 'warning',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node } = context;
-        if (!isValidNaming(node.name)) {
-          return [
-            {
-              ruleId: BUILTIN_RULE_IDS.NAMING_CONVENTION,
-              severity: 'warning',
-              message: `Directory name "${node.name}" does not follow kebab-case or camelCase conventions.`,
-              path: node.path,
-              suggestion: `Rename "${node.name}" to kebab-case (e.g. my-module) or camelCase (e.g. myModule).`,
-            },
-          ];
-        }
-        return [];
-      },
-    },
-
-    // 2. organ-no-intentmd: organ 노드에 INTENT.md가 없어야 한다
-    {
-      id: BUILTIN_RULE_IDS.ORGAN_NO_INTENTMD,
-      name: 'Organ No INTENT.md',
-      description: 'Organ nodes must not contain INTENT.md.',
-      category: 'structure',
-      severity: 'error',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node } = context;
-        if (node.type === 'organ' && node.hasIntentMd) {
-          return [
-            {
-              ruleId: BUILTIN_RULE_IDS.ORGAN_NO_INTENTMD,
-              severity: 'error',
-              message: `Organ directory "${node.name}" contains INTENT.md. Organ nodes must not have standalone documentation.`,
-              path: node.path,
-              suggestion:
-                'Remove INTENT.md or reclassify this directory as a fractal node.',
-            },
-          ];
-        }
-        return [];
-      },
-    },
-
-    // 3. index-barrel-pattern: fractal 노드의 index.ts가 순수 barrel이어야 한다
-    {
-      id: BUILTIN_RULE_IDS.INDEX_BARREL_PATTERN,
-      name: 'Index Barrel Pattern',
-      description:
-        'index.ts in fractal nodes must follow the pure barrel (re-export only) pattern.',
-      category: 'index',
-      severity: 'warning',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node } = context;
-        // fractal/hybrid 노드에만 적용, index가 있어야 함
-        if (node.type !== 'fractal' && node.type !== 'hybrid') return [];
-        if (!node.hasIndex) return [];
-
-        // metadata에 barrelPattern이 있으면 활용
-        const barrelPattern = node.metadata['barrelPattern'] as
-          | { isPureBarrel: boolean; declarationCount: number }
-          | undefined;
-
-        if (
-          barrelPattern &&
-          !barrelPattern.isPureBarrel &&
-          barrelPattern.declarationCount > 0
-        ) {
-          return [
-            {
-              ruleId: BUILTIN_RULE_IDS.INDEX_BARREL_PATTERN,
-              severity: 'warning',
-              message: `"${node.name}/index.ts" contains ${barrelPattern.declarationCount} direct declarations and does not follow the pure barrel pattern.`,
-              path: node.path,
-              suggestion:
-                'Move direct declarations into separate files and re-export them from index.ts.',
-            },
-          ];
-        }
-        return [];
-      },
-    },
-
-    // 4. module-entry-point: 모든 fractal 노드에 index.ts 또는 main.ts가 있어야 한다
-    {
-      id: BUILTIN_RULE_IDS.MODULE_ENTRY_POINT,
-      name: 'Module Entry Point',
-      description:
-        'Every fractal node must have either index.ts or main.ts as an entry point.',
-      category: 'module',
-      severity: 'warning',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node } = context;
-        if (node.type !== 'fractal' && node.type !== 'hybrid') return [];
-        if (!node.hasIndex && !node.hasMain) {
-          return [
-            {
-              ruleId: BUILTIN_RULE_IDS.MODULE_ENTRY_POINT,
-              severity: 'warning',
-              message: `Fractal module "${node.name}" does not have an entry point (index.ts or main.ts).`,
-              path: node.path,
-              suggestion:
-                'Create index.ts or main.ts and define the public API of the module there.',
-            },
-          ];
-        }
-        return [];
-      },
-    },
-
-    // 5. max-depth: 트리 깊이가 maxDepth를 초과하면 안 된다
-    {
-      id: BUILTIN_RULE_IDS.MAX_DEPTH,
-      name: 'Max Depth',
-      description:
-        'The depth of the fractal tree must not exceed the maximum allowed depth.',
-      category: 'structure',
-      severity: 'error',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node, scanOptions } = context;
-        const maxDepth = scanOptions?.maxDepth ?? DEFAULT_SCAN_OPTIONS.maxDepth;
-        if (node.depth > maxDepth) {
-          return [
-            {
-              ruleId: BUILTIN_RULE_IDS.MAX_DEPTH,
-              severity: 'error',
-              message: `The depth of "${node.name}" (${node.depth}) exceeds the maximum allowed depth (${maxDepth}).`,
-              path: node.path,
-              suggestion:
-                'Flatten the directory structure or merge related modules.',
-            },
-          ];
-        }
-        return [];
-      },
-    },
-
-    // 6. circular-dependency: 순환 의존 감지 (placeholder - 빈 배열 반환)
-    {
-      id: BUILTIN_RULE_IDS.CIRCULAR_DEPENDENCY,
-      name: 'Circular Dependency',
-      description: 'There must be no circular dependencies between modules.',
-      category: 'dependency',
-      severity: 'error',
-      enabled: true,
-      check(_context: RuleContext): RuleViolation[] {
-        // Phase 2 placeholder: 순환 의존 감지는 project-analyzer 레벨에서 수행
-        // 개별 노드 컨텍스트에서는 전체 의존 그래프가 없으므로 빈 배열 반환
-        return [];
-      },
-    },
-
-    // 7. pure-function-isolation: pure-function 노드는 부작용이 없어야 한다
-    {
-      id: BUILTIN_RULE_IDS.PURE_FUNCTION_ISOLATION,
-      name: 'Pure Function Isolation',
-      description:
-        'pure-function nodes must not import parent fractal modules.',
-      category: 'dependency',
-      severity: 'error',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node, tree } = context;
-        if (node.type !== 'pure-function') return [];
-
-        // metadata에 dependencies가 있으면 확인
-        const deps = node.metadata['dependencies'] as string[] | undefined;
-        if (!deps || deps.length === 0) return [];
-
-        const violations: RuleViolation[] = [];
-
-        for (const dep of deps) {
-          // 의존 대상이 트리에 있고 fractal 타입이면 위반
-          const depNode = tree.nodes.get(dep);
-          if (
-            depNode &&
-            (depNode.type === 'fractal' || depNode.type === 'hybrid')
-          ) {
-            violations.push({
-              ruleId: BUILTIN_RULE_IDS.PURE_FUNCTION_ISOLATION,
-              severity: 'error',
-              message: `Pure-function node "${node.name}" depends on fractal module "${depNode.name}".`,
-              path: node.path,
-              suggestion: `Move the dependency under "${depNode.name}" as an organ node or remove the dependency.`,
-            });
-          }
-        }
-
-        return violations;
-      },
-    },
-
-    // 8. zero-peer-file: 카테고리 기반 예외 시스템 (static + eponymous + framework)
-    {
-      id: BUILTIN_RULE_IDS.ZERO_PEER_FILE,
-      name: 'Zero Peer File',
-      description:
-        'Fractal roots must not contain standalone peer files beyond the allowed categories (index/main, documentation, eponymous, framework reserved).',
-      category: 'structure',
-      severity: 'warning',
-      enabled: true,
-      check(context: RuleContext): RuleViolation[] {
-        const { node } = context;
-        if (node.type !== 'fractal' && node.type !== 'hybrid') return [];
-
-        const peerFiles = node.metadata['peerFiles'] as string[] | undefined;
-        if (!peerFiles || peerFiles.length === 0) return [];
-
-        // Compose the full allowed set for THIS node
-        const allowed = new Set(ALLOWED_FRACTAL_ROOT_FILES);
-
-        // Category: eponymous file (max 1, auto-detected by scanProject)
-        const eponymous = node.metadata['eponymousFile'] as
-          | string
-          | null
-          | undefined;
-        if (eponymous) allowed.add(eponymous);
-
-        // Category: framework reserved files (auto-detected from package.json)
-        const fwFiles = node.metadata['frameworkReservedFiles'] as
-          | string[]
-          | undefined;
-        if (fwFiles) for (const f of fwFiles) allowed.add(f);
-
-        // Category: additional-allowed from .filid/config.json
-        //   string entry  → allowed everywhere (backward-compat).
-        //   object entry  → allowed only when entry.paths glob matches node.path.
-        if (additionalAllowed) {
-          for (const entry of additionalAllowed) {
-            if (typeof entry === 'string') {
-              allowed.add(entry);
-              continue;
-            }
-            if (entry.paths && !isExempt(node, entry.paths)) continue;
-            allowed.add(entry.basename);
-          }
-        }
-
-        const disallowed = peerFiles.filter((f) => !allowed.has(f));
-        if (disallowed.length === 0) return [];
-
-        return disallowed.map((file) => ({
-          ruleId: BUILTIN_RULE_IDS.ZERO_PEER_FILE,
-          severity: 'warning' as const,
-          message: `Fractal root "${node.name}" contains peer file "${file}" not in any allowed category. Promote it to a sub-fractal directory.`,
-          path: node.path,
-          suggestion: `Create a subdirectory for "${file}" or add it to .filid/config.json additional-allowed if it belongs at the root.`,
-        }));
-      },
-    },
-  ];
-  return overrides ? applyOverrides(rules, overrides) : rules;
-}
-
-/**
- * 활성화된 규칙만 필터링하여 반환한다.
- *
- * @param rules - 전체 규칙 목록
- * @returns enabled가 true인 규칙 목록
- */
-export function getActiveRules(rules: Rule[]): Rule[] {
-  return rules.filter((r) => r.enabled);
-}
-
-/**
- * 단일 규칙을 단일 노드에 적용한다.
- *
- * @param rule - 적용할 규칙
- * @param context - 규칙 컨텍스트
- * @returns 위반 목록 (없으면 빈 배열)
- */
-export function evaluateRule(
-  rule: Rule,
-  context: RuleContext,
-): RuleViolation[] {
-  if (!rule.enabled) return [];
-  try {
-    return rule.check(context);
-  } catch {
-    return [];
-  }
-}
-
-/**
- * FractalTree의 모든 노드에 대해 활성화된 규칙을 평가한다.
- *
- * @param tree - 검증할 프랙탈 트리
- * @param rules - 평가할 규칙 목록 (생략 시 내장 규칙 사용)
- * @param options - 스캔 옵션 (maxDepth 등)
- * @returns 전체 평가 결과
- */
-export function evaluateRules(
-  tree: FractalTree,
-  rules?: Rule[],
-  options?: ScanOptions,
-): RuleEvaluationResult {
-  const start = Date.now();
-  const activeRules = getActiveRules(rules ?? loadBuiltinRules());
-  const violations: RuleViolation[] = [];
-  let passed = 0;
-  let failed = 0;
-  let skipped = 0;
-
-  for (const [, node] of tree.nodes) {
-    const context: RuleContext = { node, tree, scanOptions: options };
-    for (const rule of activeRules) {
-      if (!rule.enabled) {
-        skipped++;
-        continue;
-      }
-      const nodeViolations = evaluateRule(rule, context);
-      if (nodeViolations.length === 0) {
-        passed++;
-      } else {
-        failed++;
-        violations.push(...nodeViolations);
-      }
-    }
-  }
-
-  return {
-    violations,
-    passed,
-    failed,
-    skipped,
-    duration: Date.now() - start,
-  };
-}
+export { applyOverrides } from './apply-overrides.js';
+export { evaluateRule } from './evaluate-rule.js';
+export { evaluateRules } from './evaluate-rules.js';
+export { getActiveRules } from './get-active-rules.js';
+export { loadBuiltinRules } from './load-builtin-rules.js';
